move append and rename _utils
