+++ chirho/dynamical/handlers/trace.py
@@ -1,47 +1,13 @@
-import functools
 from typing import Generic, TypeVar
 
 import pyro
 import torch
 
+from chirho.dynamical.internals._utils import append
 from chirho.dynamical.internals.backend import simulate_trajectory
 from chirho.dynamical.ops import Trajectory
 
 T = TypeVar("T")
-
-
-@functools.singledispatch
-def append(fst, rest: T) -> T:
-    raise NotImplementedError(f"append not implemented for type {type(fst)}.")
-
-
-@append.register(Trajectory)
-def append_trajectory(traj1: Trajectory[T], traj2: Trajectory[T]) -> Trajectory[T]:
-    if len(traj1.keys) == 0:
-        return traj2
-
-    if len(traj2.keys) == 0:
-        return traj1
-
-    if traj1.keys != traj2.keys:
-        raise ValueError(
-            f"Trajectories must have the same keys to be appended, but got {traj1.keys} and {traj2.keys}."
-        )
-
-    result: Trajectory[T] = Trajectory()
-    for k in traj1.keys:
-        setattr(result, k, append(getattr(traj1, k), getattr(traj2, k)))
-
-    return result
-
-
-@append.register(torch.Tensor)
-def append_tensor(prev_v: torch.Tensor, curr_v: torch.Tensor) -> torch.Tensor:
-    time_dim = -1  # TODO generalize to nontrivial event_shape
-    batch_shape = torch.broadcast_shapes(prev_v.shape[:-1], curr_v.shape[:-1])
-    prev_v = prev_v.expand(*batch_shape, *prev_v.shape[-1:])
-    curr_v = curr_v.expand(*batch_shape, *curr_v.shape[-1:])
-    return torch.cat([prev_v, curr_v], dim=time_dim)
 
 
 class DynamicTrace(Generic[T], pyro.poutine.messenger.Messenger):
