+++ chirho/dynamical/internals/__init__.py
@@ -1,4 +1,4 @@
 # Include only imports that are needed for registering dispatches.
 
-from . import _patterns  # noqa: F401
+from . import _utils  # noqa: F401
 from . import solver  # noqa: F401
